Modify README, and create subfolder for statement_insurer
