+++ lib/safe-pg-migrations/base.rb
@@ -5,7 +5,7 @@ require 'safe-pg-migrations/helpers/satisfied_helper'
 require 'safe-pg-migrations/helpers/index_helper'
 require 'safe-pg-migrations/plugins/verbose_sql_logger'
 require 'safe-pg-migrations/plugins/blocking_activity_logger'
-require 'safe-pg-migrations/plugins/statement_insurer_add_column'
+require 'safe-pg-migrations/plugins/statement_insurer/add_column'
 require 'safe-pg-migrations/plugins/statement_insurer'
 require 'safe-pg-migrations/plugins/statement_retrier'
 require 'safe-pg-migrations/plugins/idempotent_statements'
